pyproject.toml
[project]
name = "streaming-memory"
version = "0.1.0"
description = "Hebbian-inspired dynamic memory system for LLM context injection"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.0",
    "pydantic>=2.0",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "modal>=1.3.0.post1",
    "backoff>=2.2.1",
    "sentence-transformers>=2.2.0",
    "transformers>=4.40.0",
    "torch>=2.0.0",
    "sentencepiece>=0.1.99",
    "tiktoken>=0.12.0",
    "websockets>=12.0",
    "deepgram-sdk>=3.0",
]

[project.optional-dependencies]
modal = [
    "modal>=0.64",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["streaming_memory"]

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "W"]
ignore = ["E501"]
